wip: step3
I'm not content with the distinction between various primitive types.

Currently, python objects are exposed to the interpreter. This leads
to problems with the new symbol type. I need to make it hashable to
fit as key into a python dict, which is used for namespaces. It might
be clearer to do distinction between primitive types with a mal-level
exposed compound type, like a vector.